+++ CHANGELOG.md
@@ -6,9 +6,17 @@ and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added ability to sort events in web UI.
+
 ### Changed
 - Changed the maximum number of open file descriptors on a system to from 1024
 (default) to 65535.
+- Increased the default etcd size limit from 2GB to 4GB.
+- Move Hooks and Silenced out of Event and into Check.
+
+### Added
+- Support for managing mutators via sensuctl.
 
 ## [2.0.0-nightly.1] - 2018-03-07
 ### Added
@@ -31,6 +39,8 @@ set when importing legacy settings.
 of a successful check.
 - Added GraphQL explorer to web UI.
 - Added check occurrences and occurrences_watermark attributes from Sensu 1.x.
+- Added issue template for GitHub.
+- Added custom functions to evaluate a unix timestamp in govaluate.
 
 ### Changed
 - Refactor Check data structure to not depend on CheckConfig. This is a breaking
@@ -56,6 +66,7 @@ code and severity (ex. 0, ok).
 - Updated the sensuctl guidelines.
 - Changed travis badge to use travis-ci.org in README.md.
 - Govaluate's modifier tokens can now be optionally forbidden.
+- Increase the stack size on Travis CI.
 - Refactor store, queue and ring interfaces, and daemon I/O details.
 
 ### Fixed